Refactor AbstractBlockHighlighter
